Fixed compile issue on jdk6: Caused by: idobata4j-core/src/main/java/com/uphyca/idobata/http/UrlConnectionClient.java:95: setFixedLengthStreamingMode(int) in java.net.HttpURLConnection cannot be applied to (long)

+++ idobata4j-core/src/main/java/com/uphyca/idobata/http/UrlConnectionClient.java
@@ -92,7 +92,7 @@ public class UrlConnectionClient implements Client {
         connection.addRequestProperty("Content-Type", body.mimeType());
         long length = body.length();
         if (length != -1) {
-            connection.setFixedLengthStreamingMode(length);
+            connection.setFixedLengthStreamingMode((int)length);
             connection.addRequestProperty("Content-Length", String.valueOf(length));
         } else {
             connection.setChunkedStreamingMode(CHUNK_SIZE);
